backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.5
python-dotenv>=1.0
python-multipart>=0.0.6
pandas>=2.0
numpy>=1.24
torch>=2.0
sentence-transformers>=2.2
ultralytics>=8.0
google-generativeai>=0.7
Pillow>=10.0
unidecode>=1.3

